fdj/requirements.txt
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.3
cryptography==36.0.0
cycler==0.11.0
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.29.1
greenlet==1.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numexpr==2.8.1
numpy==1.21.5
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
pip==21.2.4
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
setuptools==58.0.4
six==1.16.0
SQLAlchemy==1.4.27
threadpoolctl==2.2.0
urllib3==1.26.7
Werkzeug==2.0.2
wheel==0.37.0
